% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qexp_arma}
\alias{qexp_arma}
\title{Quantile function for exponential}
\usage{
qexp_arma(p, lambda = 2)
}
\arguments{
\item{p}{Vector of Probabilities.}

\item{lambda}{Double of lambda.}
}
\value{
A vector of quantiles
}
\description{
Compute the quantile given a probability
}
\examples{
# Compute one quantile
qexp_arma(0.6, 2)

# Compute two quantiles
qexp_arma(c(0.5, 0.2), 3)
}
