% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rexp_arma}
\alias{rexp_arma}
\title{Simulate samples from an Exponential Distribution}
\usage{
rexp_arma(n, lambda = 2, seed = 183L)
}
\arguments{
\item{n}{Number of observations to generate}

\item{lambda}{Rate parameter}

\item{seed}{Starting value for the Random Number Generator (RNG).}
}
\value{
A vector of random numbers from an exponential distribution.
}
\description{
Generates random values from an exponential distribution.
}
\examples{
# Generate a random number
rexp_arma(1, lambda = 2)

# Generate two random numbers
rexp_arma(2, lambda = 9, seed = 8235)
}
