% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runif_arma}
\alias{runif_arma}
\title{Simulate samples from a Uniform Distribution}
\usage{
runif_arma(n, start = 0, end = 1, seed = 183L)
}
\arguments{
\item{n}{Number of observations to generate}

\item{start}{Beginning range of values}

\item{end}{End range of values}

\item{seed}{Starting value for the Random Number Generator (RNG).}
}
\value{
A vector of random numbers from the Uniform distribution.
}
\description{
Generates random values from the C++11 Uniform Distribution.
}
\examples{
# Generate a random number
runif_arma(1, start = 2, end = 20)

# Generate two random numbers
runif_arma(2, start = 5, end = 30)
}
