% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runif_cpp}
\alias{runif_cpp}
\title{Simulate Samples from a Uniform Distribution}
\usage{
runif_cpp(n, start = 0, end = 1, seed = 183L)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{start}{Lower bound of the distribution.}

\item{end}{Upper bound of the distribution.}

\item{seed}{Seed for the random number generator.}
}
\value{
A \code{numeric} vector of \code{n} draws from the Uniform(\code{start}, \code{end})
distribution.
}
\description{
Generates random values from a uniform distribution using the C++ standard
library's \verb{<random>} facilities, which have been part of standard C++ since
C++11.
}
\examples{
# Generate one draw
runif_cpp(1, start = 2, end = 20)

# Generate two draws
runif_cpp(2, start = 5, end = 30)
}
