% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r11-package.r
\docType{package}
\name{r11-package}
\alias{r11}
\alias{r11-package}
\title{r11: Compiling code with C++11}
\description{
Compile code inside a package using C++11 instead of the default
    language specification of C++98.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/r-pkg-examples/rcpp-cpp11-usage}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
